Excclude conditional check, more refactoring
