--- src/NetAnalyzers/Core/Microsoft.NetCore.Analyzers/InteropServices/MockOsPlatformAttribute.cs
@@ -1,124 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-namespace System.Runtime.Versioning
-{
-    public abstract class OSPlatformAttribute : Attribute
-    {
-        private protected OSPlatformAttribute(string platformName)
-        {
-            PlatformName = platformName;
-        }
-
-        public string PlatformName { get; }
-    }
-
-    [AttributeUsage(AttributeTargets.Assembly,
-                    AllowMultiple = false, Inherited = false)]
-    public sealed class TargetPlatformAttribute : OSPlatformAttribute
-    {
-        public TargetPlatformAttribute(string platformName) : base(platformName)
-        { }
-    }
-
-    [AttributeUsage(AttributeTargets.Assembly |
-                    AttributeTargets.Class |
-                    AttributeTargets.Constructor |
-                    AttributeTargets.Event |
-                    AttributeTargets.Method |
-                    AttributeTargets.Module |
-                    AttributeTargets.Property |
-                    AttributeTargets.Struct |
-                    AttributeTargets.Field,
-                    AllowMultiple = true, Inherited = false)]
-    public sealed class MinimumOSPlatformAttribute : OSPlatformAttribute
-    {
-        public MinimumOSPlatformAttribute(string platformName) : base(platformName)
-        { }
-    }
-
-    [AttributeUsage(AttributeTargets.Assembly |
-                    AttributeTargets.Class |
-                    AttributeTargets.Constructor |
-                    AttributeTargets.Event |
-                    AttributeTargets.Method |
-                    AttributeTargets.Module |
-                    AttributeTargets.Property |
-                    AttributeTargets.Struct |
-                    AttributeTargets.Field,
-                    AllowMultiple = true, Inherited = false)]
-    public sealed class RemovedInOSPlatformAttribute : OSPlatformAttribute
-    {
-        public RemovedInOSPlatformAttribute(string platformName) : base(platformName)
-        { }
-    }
-
-    [AttributeUsage(AttributeTargets.Assembly |
-                    AttributeTargets.Class |
-                    AttributeTargets.Constructor |
-                    AttributeTargets.Event |
-                    AttributeTargets.Method |
-                    AttributeTargets.Module |
-                    AttributeTargets.Property |
-                    AttributeTargets.Struct |
-                    AttributeTargets.Field,
-                    AllowMultiple = true, Inherited = false)]
-    public sealed class ObsoletedInOSPlatformAttribute : OSPlatformAttribute
-    {
-        public ObsoletedInOSPlatformAttribute(string platformName) : base(platformName)
-        { }
-        public ObsoletedInOSPlatformAttribute(string platformName, string message) : base(platformName)
-        {
-            Message = message;
-        }
-        public string? Message { get; }
-        public string? Url { get; set; }
-    }
-}
-
-namespace System.Runtime.InteropServices
-{
-    public static class RuntimeInformationHelper
-    {
-#pragma warning disable CA1801, IDE0060 // Review unused parameters
-        public static bool IsOSPlatformOrLater(OSPlatform osPlatform, int major)
-        {
-            return RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformOrLater(OSPlatform osPlatform, int major, int minor)
-        {
-            return RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformOrLater(OSPlatform osPlatform, int major, int minor, int build)
-        {
-            return RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformOrLater(OSPlatform osPlatform, int major, int minor, int build, int revision)
-        {
-            return RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformEarlierThan(OSPlatform osPlatform, int major)
-        {
-            return !RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformEarlierThan(OSPlatform osPlatform, int major, int minor)
-        {
-            return !RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformEarlierThan(OSPlatform osPlatform, int major, int minor, int build)
-        {
-            return !RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-
-        public static bool IsOSPlatformEarlierThan(OSPlatform osPlatform, int major, int minor, int build, int revision)
-        {
-            return !RuntimeInformation.IsOSPlatform(osPlatform);
-        }
-#pragma warning restore CA1801, IDE0060 // Review unused parameters
-    }
-}
